Version 3
This version has several fixes, plus several new features:

1) BibSearch opens now a new window, a syntax file for the results is provided, which is based on syntax file for bib files (see :h atp-bibsearch-highlight)
2) TOC command, which displays in seperate window interactive table of contents (wich the same numbering as in the output file),
     I also provide a syntax file for TOC files
3) Lables command which dispalys in a seprate interactiv window lables found in your file

Windows openned by TOC and Labels commands have functions which make it easy to copy the label (of chapter or of the label) to a registar or directly to you file, or to move to the label, or to echo the line with the label and to split the window of your tex file and one of them move to the label, to see its contens.

If you open table of contents or the list of labels cursor position is set to the chapter (label) which you are currently editing.

see :h atp-TOC() and :h atp-Labels()

Best regards and happy texing :)
Marcin Szamotulski
4) I provide a color file, called coot-beauty-256 (it needs support for 256 colors).
4) Status line now shows the current chapter/section of the edited tex file.